api/requirements.txt
flask
Flask-SQLAlchemy
psycopg2-binary
pydantic
python-dotenv